project-b-22/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-allauth==0.57.0
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
filelock==3.12.4
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.2
platformdirs==3.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
virtualenv==20.24.5
whitenoise==6.6.0
Pillow==10.1.0